refactor(web-serial-rxjs)!: expose only SerialSession as v2 public API
Remove the legacy `SerialClient` / `ShellClient` / browser / io / filters
surface from the library public API as the final step of the v2 redesign
(#199). The only remaining public exports are `createSerialSession`,
`SerialSession`, `SerialSessionOptions`, `SerialSessionState`,
`SerialError`, and `SerialErrorCode`.

- Delete `src/{client,shell,browser,io,filters,lib,types}`
- Inline the default session options and `SerialPortFilter`-friendly
  request-options builder under `src/session/internal/`
- Remove the `./shell` package export entry from package.json

BREAKING CHANGE: `createSerialClient`, `createShellClient`, the
top-level `isBrowserSupported`, `SerialClient`, `ShellClient`,
`observableToWritable`, `readableToObservable`, and `buildRequestOptions`
are no longer exported. Migrate to `createSerialSession` and use
`session.isBrowserSupported()`; see the upcoming `docs/MIGRATION_V2.md`.

Refs #199

Made-with: Cursor
