+++ packages/web-serial-rxjs/src/index.ts
@@ -3,59 +3,52 @@
  *
  * # web-serial-rxjs
  *
- * A TypeScript library that provides a reactive RxJS-based wrapper for the Web Serial API,
- * enabling easy serial port communication in web applications.
+ * A TypeScript library that wraps the Web Serial API with a minimal,
+ * RxJS-based session surface.
  *
- * ## Features
+ * ## Public API (v2)
  *
- * - **RxJS-based reactive API**: Leverage the power of RxJS Observables for reactive serial port communication
- * - **TypeScript support**: Full TypeScript type definitions included
- * - **Browser detection**: Built-in browser support detection and error handling
- * - **Error handling**: Comprehensive error handling with custom error classes and error codes
- * - **Framework agnostic**: Works with any JavaScript/TypeScript framework or vanilla JavaScript
+ * The v2 public API intentionally exposes a single, session-oriented surface
+ * so that apps (Angular, Vue, React, Svelte, vanilla JS/TS) can drive their
+ * UI entirely from `state$` + `receive$` + `errors$` without rebuilding any
+ * state, read loops, or write queues themselves.
+ *
+ * - {@link createSerialSession} - factory for a {@link SerialSession}
+ * - {@link SerialSession} - the runtime interface
+ * - {@link SerialSessionOptions} - connection options
+ * - {@link SerialSessionState} - `state$` payload type
+ * - {@link SerialError} / {@link SerialErrorCode} - unified error surface
  *
  * ## Browser Support
  *
- * The Web Serial API is currently only supported in Chromium-based browsers:
+ * The Web Serial API is only available in Chromium-based browsers:
+ *
  * - Chrome 89+
  * - Edge 89+
  * - Opera 75+
  *
+ * Use {@link SerialSession.isBrowserSupported} for a synchronous feature
+ * check before calling {@link SerialSession.connect$}.
+ *
  * @example
  * ```typescript
- * import { createSerialClient, isBrowserSupported } from '@gurezo/web-serial-rxjs';
+ * import { createSerialSession } from '@gurezo/web-serial-rxjs';
  *
- * // Check browser support
- * if (!isBrowserSupported()) {
+ * const session = createSerialSession({ baudRate: 115200 });
+ *
+ * if (!session.isBrowserSupported()) {
  *   console.error('Web Serial API is not supported in this browser');
- *   return;
+ * } else {
+ *   session.state$.subscribe((state) => console.log('state:', state));
+ *   session.receive$.subscribe((chunk) => console.log('rx:', chunk));
+ *   session.errors$.subscribe((error) => console.error('err:', error));
+ *
+ *   session.connect$().subscribe();
+ *   session.send$('hello\r\n').subscribe();
  * }
- *
- * // Create a serial client
- * const client = createSerialClient({ baudRate: 9600 });
- *
- * // Connect to a serial port
- * client.connect().subscribe({
- *   next: () => console.log('Connected!'),
- *   error: (error) => console.error('Connection error:', error),
- * });
  * ```
  */
 
-// Core API (recommended)
-export { createSerialClient } from './client';
-
-// Type exports
-export type {
-  CommandResult,
-  SerialClient,
-  SerialCommandOptions,
-  SerialRequest,
-  SerialState,
-  SerialSupport,
-} from './client';
-
-// v2 SerialSession API (additive, stabilizing under #199)
 export { createSerialSession } from './session';
 export type {
   SerialSession,
@@ -63,36 +56,5 @@ export type {
   SerialSessionState,
 } from './session';
 
-// Advanced API
-export { createShellClient } from './shell';
-export type { ShellClient, ShellClientOptions, ShellExecResult } from './shell';
-
-// Error exports
 export { SerialError } from './errors/serial-error';
 export { SerialErrorCode } from './errors/serial-error-code';
-
-// Options type exports
-export type { SerialClientOptions } from './types/options';
-
-// Browser support exports
-export {
-  BrowserType,
-  detectBrowserType,
-  hasWebSerialSupport,
-  isChromiumBased,
-} from './browser/browser-detection';
-export {
-  checkBrowserSupport,
-  isBrowserSupported,
-} from './browser/browser-support';
-
-// Compatibility/utility exports
-// These are low-level helpers and are not required for standard v1 usage.
-export {
-  observableToWritable,
-  subscribeToWritable,
-} from './io/observable-to-writable';
-export { readableToObservable } from './io/readable-to-observable';
-
-// Filter exports
-export { buildRequestOptions } from './filters/build-request-options';
