--- packages/web-serial-rxjs/src/lib/web-serial-rxjs.spec.ts
@@ -1,7 +1,0 @@
-import { webSerialRxjs } from './web-serial-rxjs';
-
-describe('webSerialRxjs', () => {
-  it('should work', () => {
-    expect(webSerialRxjs()).toEqual('web-serial-rxjs');
-  });
-});
