--- packages/web-serial-rxjs/src/lib/web-serial-rxjs.ts
@@ -1,8 +1,0 @@
-/**
- * Utility function that returns the library name.
- *
- * @returns The library name as a string: 'web-serial-rxjs'
- */
-export function webSerialRxjs(): string {
-  return 'web-serial-rxjs';
-}
