--- packages/web-serial-rxjs/src/shell/index.ts
@@ -1,6 +1,0 @@
-export {
-  createShellClient,
-  type ShellClient,
-  type ShellClientOptions,
-  type ShellExecResult,
-} from './create-shell-client';
